Start splitting facet-derive into parse & emit
Closes #175
