+++ facet-derive-emit/README.md.in
@@ -1,0 +1,1 @@
+Emits Rust code for [facet-derive](https://github.com/facet-rs/facet/tree/main/facet-derive)
