+++ facet-derive-parse/README.md.in
@@ -1,0 +1,1 @@
+Parses Rust syntax for [facet-derive](https://github.com/facet-rs/facet/tree/main/facet-derive)
